requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.4
cssselect==1.2.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
greenlet==2.0.2
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
jedi==0.19.0
jieba3k==0.35.1
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
lxml==4.9.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
newspaper3k==0.2.8
nltk==3.8.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
playwright==1.36.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyee==9.0.4
Pygments==2.16.1
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.0
redis==4.6.0
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
tinysegmenter==0.3
tldextract==3.4.4
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
